Change in data of things for test

+++ database/migrations/2022_05_13_220658_create_affiliates_table.php
@@ -18,7 +18,11 @@ return new class extends Migration
             $table->foreignId('user_id');
             $table->foreignId('merchant_id');
             // TODO: Replace me with a brief explanation of why floats aren't the correct data type, and replace with the correct data type.
-            $table->float('commission_rate');
+            // Because of possible accuracy and rounding difficulties, floats are not the appropriate data type to represent monetary values such as commission rates or discounts. 
+            // Floats are binary representations of numbers that might cause mistakes when conducting computations with decimal values.
+            // To correctly represent monetary amounts, it is advised to utilise decimal data types rather than floats. 
+            // In Laravel, you may utilise the decimal method to define these columns in your migration file.
+            $table->decimal('commission_rate', 2, 2);
             $table->string('discount_code');
             $table->timestamps();
         });
